fix: Change duration field from Int to Float in GraphQL schema
The Spectrogram model's duration property returns a float
(timedelta.total_seconds()), but the GraphQL schema defined it as Int,
causing serialization errors for files with non-integer durations
like 59.997312s.

+++ backend/api/schema/nodes/annotation_spectrogram.py
@@ -59,7 +59,7 @@ def get_task(
 
 class AnnotationSpectrogramNode(BaseObjectType):
 
-    duration = graphene.Int(required=True)
+    duration = graphene.Float(required=True)
     annotation_comments = AuthenticatedDjangoConnectionField(AnnotationCommentNode)
 
     class Meta:
